Deleting Facade classes ("SocketsLoggin.cs")
As discussed, facades classes are not necessary anymore, since now there
are no common event methods between eventsources. Enter, Exit and
CriticalEvents will be exclusive to NetEventSource. This will be the
only event methods present in NetEventSource (previously
WebEventSource). Older event methods that were in WebEventSource will be
moved to a new SecurityEventSource since Security is the only consumer
of those events.

In this change:

-Adding the proposed Enter method in NetEventSource.

-Adding an example of Accepted event method in the SocketsEventSource
class.

-Adding HashLogging class with some common hashing methods.

-Removing facades.

-Changing enums used as event ids for const int types.